fix(themes): use fitting desktop colors for remaining themes
Use a fitting hand-picked desktop background color that matches the
existing colors and spirit of the theme.

For theSixtiesUSA and travel, the two themes that originate from the
Windows 98 Plus! pack, colors based on the original wallpaper were
chosen. The original themes themselves specify full black as the desktop
background, supposedly with the assumption that the included wallpaper
would be used, making the background color irrelevant.

coldGray and lilacRoseDark appear similar to themes published at
Mochinet (https://www.mochinet.com/Themes/), but they do not appear to
be the same themes. New colors were chosen instead of taking from the
Mochinet themes since the Mochinet themes are low contrast, whereas the
React95 themes are not.